+++ out/extension.js
@@ -1,15 +1,10 @@
-"use strict"
+"use strict";
 
-const vscode = require("vscode")
+const vscode = require("vscode");
 
-const {
-	Theme
-} = require("../src/_make-theme/_make-theme")
+const { Theme } = require("../src/_make-theme/_make-theme");
 
-const {
-	markdownFile,
-	themeMenuName
-} = require("../data/theme-names")
+const { markdownFile, themeMenuName } = require("../data/theme-names");
 
 const {
 	info,
@@ -17,21 +12,21 @@ const {
 	reloadMessage,
 	updateMessage,
 	website,
-} = require("../src/vscode/info-message")
+} = require("../src/vscode/info-message");
 
 const {
 	favoriteConfig,
 	saveSettings,
 	themeConfig,
 	freshInstall,
-} = require("../src/vscode/user-preferences")
+} = require("../src/vscode/user-preferences");
 
 const {
 	applyChanges,
 	applyChangesFalse,
 	favoriteOFF,
 	getConfig,
-} = require("../src/vscode/menu-settings")
+} = require("../src/vscode/menu-settings");
 
 const {
 	createFile,
@@ -39,7 +34,7 @@ const {
 	dataFileName,
 	fileMessage,
 	noFile,
-} = require("../src/vscode/theme-control")
+} = require("../src/vscode/theme-control");
 
 // ! for tests
 // info("FIRST")
@@ -48,29 +43,29 @@ const {
 
 function activate() {
 	if (noFile) {
-		createFile(dataDir, dataFileName, markdownFile, fileMessage)
+		createFile(dataDir, dataFileName, markdownFile, fileMessage);
 		if (freshInstall) {
-			info(installMessage, "More Info", website)
-			saveSettings(getConfig())
+			info(installMessage, "More Info", website);
+			saveSettings(getConfig());
 		} else {
-			info(updateMessage, "OK")
-			new Theme(themeConfig()).allThemes()
-			new Theme(favoriteConfig()).favoriteTheme()
+			info(updateMessage, "OK");
+			new Theme(themeConfig()).allThemes();
+			new Theme(favoriteConfig()).favoriteTheme();
 		}
 	}
 	vscode.workspace.onDidChangeConfiguration((event) => {
 		if (event.affectsConfiguration(themeMenuName)) {
 			if (applyChanges()) {
-				const config = getConfig()
-				new Theme(config).allThemes()
-				new Theme(config).favoriteTheme()
-				applyChangesFalse()
-				favoriteOFF()
-				info(reloadMessage, "Reload", true)
-				saveSettings(config)
+				const config = getConfig();
+				new Theme(config).allThemes();
+				new Theme(config).favoriteTheme();
+				applyChangesFalse();
+				favoriteOFF();
+				info(reloadMessage, "Reload", true);
+				saveSettings(config);
 			}
 		}
-	})
+	});
 }
 
 function deactivate() {}
@@ -78,4 +73,4 @@ function deactivate() {}
 module.exports = {
 	activate,
 	deactivate,
-}
+};