+++ src/_make-theme/make-syntax.js
@@ -1,55 +1,47 @@
-"use strict"
+"use strict";
 
-const {
-	nameOptions
-} = require("../../data/theme-names")
+const { nameOptions } = require("../../data/theme-names");
 
-const {
-	makeTextmate,
-	makeSemantic
-} = require("../syntax/_syntax-base")
+const { makeTextmate, makeSemantic } = require("../syntax/_syntax-base");
 
 const {
 	syntaxColors,
-	colorSchemes,
 	commentColors,
-} = require("../syntax/_color-palette")
+	schemes,
+} = require("../syntax/_color-palette");
 
-const {
-	commentTokens
-} = require("../syntax/syntax-extra")
+const { commentTokens } = require("../syntax/syntax-extra");
 
 class Syntax {
-	constructor(fontStyle, comments) {
+	constructor(fontStyle, comments, syntaxBrightness) {
 		Object.values(nameOptions).map((name) => {
 			this[name] = {
 				semanticHighlighting: true,
 				semanticTokenColors: makeSemantic(
 					fontStyle,
-					colorSchemes,
+					schemes(nameOptions, syntaxColors[syntaxBrightness]),
 					name
 				),
 				tokenColors: [
-					...commentTokens(comments,
-						commentColors),
+					...commentTokens(comments, commentColors[syntaxBrightness]),
 					...makeTextmate(
 						fontStyle,
-						syntaxColors,
+						syntaxColors[syntaxBrightness],
 						nameOptions,
-						colorSchemes,
+						schemes(nameOptions, syntaxColors[syntaxBrightness]),
 						name
 					),
 				],
-			}
-		})
+			};
+		});
 	}
 }
 
 module.exports = {
 	Syntax,
-}
+};
 
-// console.log(new Syntax("Italic", "Dark Violet"))
-// console.log(new Syntax("Italic", "Dark Magenta"))
-// console.log(new Syntax("Italic", "Dark Forest"))
-// console.log(new Syntax("Italic", "Dark Grey"))
+// console.log(new Syntax("Italic", "Dark Violet", "Normal"));
+// console.log(new Syntax("Italic", "Dark Grey", "Normal"));))
+// console.log(new Syntax("Italic", "Dark Magenta", "Dimmed"));))
+// console.log(new Syntax("Italic", "Dark Forest", "Dimmed"));))