+++ src/syntax/_color-palette.js
@@ -1,132 +1,70 @@
-"use strict"
-
-const {
-	nameOptions
-} = require("../../data/theme-names")
-
-// const {
-// 	getConfig
-// } = require("../vscode/menu-settings")
+"use strict";
 
 const syntaxColors = {
-	pink: "#FF9EE7",
-	magenta: "#D74BBE",
-	purple: "#AB80FF",
-	blue: "#50B6FF",
-	cyan: "#4FE0E0",
-	green: "#2FDE73",
-	yellow: "#E6E667",
-	white: "#D3D1EB",
-	red: "#E6457B",
-	brown: "#99611F",
-}
+	Normal: {
+		pink: "#FF9EE7",
+		magenta: "#D74BBE",
+		purple: "#AB80FF",
+		blue: "#50B6FF",
+		cyan: "#4FE0E0",
+		green: "#2FDE73",
+		yellow: "#E6E667",
+		white: "#D3D1EB",
+		red: "#E6457B",
+		brown: "#99611F",
+	},
+	Dimmed: {
+		pink: "#E18BCB",
+		magenta: "#BD42AA",
+		purple: "#8E6CD7",
+		blue: "#4197D5",
+		cyan: "#41BAB9",
+		green: "#29BD62",
+		yellow: "#CBCB5B",
+		white: "#BBBAD2",
+		red: "#CC3D69",
+		brown: "#82511A",
+	},
+};
 
-// const syntaxDimmed = (colors, dimmed) => {
-// 	if (dimmed === "dimmed") {
-// 		return colors.dimmed
-// 	} else {
-// 		return colors.normal
-// 	}
-// }
-
-// const syntaxColors = syntaxDimmed(colors, dimmed)
-
-// normal
-// const syntaxColors = {
-// 	pink: "#FF9EE7",
-// 	magenta: "#D74BBE",
-// 	purple: "#AB80FF",
-// 	blue: "#50B6FF",
-// 	cyan: "#4FE0E0",
-// 	green: "#2FDE73",
-// 	yellow: "#E6E667",
-// 	white: "#D3D1EB",
-// 	red: "#E6457B",
-// 	brown: "#99611F",
-// }
-
-// dimmed
-// const syntaxColors = {
-// 	pink: "#CF80BB",
-// 	magenta: "#B13D9C",
-// 	purple: "#8A68CF",
-// 	blue: "#4093CF",
-// 	cyan: "#3FB5B5",
-// 	green: "#26B45D",
-// 	yellow: "#BABA54",
-// 	white: "#ABA9BE",
-// 	red: "#BA3864",
-// 	brown: "#7C4F19",
-// }
-
-// extra dimmed
-// const syntaxColors = {
-// 	pink: "#C47AB2",
-// 	magenta: "#A73A95",
-// 	purple: "#8363C5",
-// 	blue: "#3E8CC4",
-// 	cyan: "#3DACAC",
-// 	green: "#23AB59",
-// 	yellow: "#B1B14F",
-// 	white: "#A2A1B5",
-// 	red: "#B1355F",
-// 	brown: "#764B17",
-// }
-
-// atom
-// const syntaxColors = {
-// 	pink: "#E5C07B",
-// 	magenta: "#E06C75",
-// 	purple: "#C678DD",
-// 	blue: "#98C379",
-// 	cyan: "#56B6C2",
-// 	green: "#E5C07B",
-// 	yellow: "#61AFEF",
-// 	white: "#ABB2BF",
-// 	red: "#E6457B",
-// 	brown: "#99611F",
-// }
-
-// const syntaxColors = {
-// 	normal: {
-// 		pink: "#FF9EE7",
-// 		magenta: "#D74BBE",
-// 		purple: "#AB80FF",
-// 		blue: "#50B6FF",
-// 		cyan: "#4FE0E0",
-// 		green: "#2FDE73",
-// 		yellow: "#E6E667",
-// 		white: "#D3D1EB",
-// 		red: "#E6457B",
-// 		brown: "#99611F",
-// 	},
-// 	dimmed: {
-// 		pink: "#FF9EE7CC",
-// 		magenta: "#D74BBECC",
-// 		purple: "#AB80FFCC",
-// 		blue: "#50B6FFCC",
-// 		cyan: "#4FE0E0CC",
-// 		green: "#2FDE73CC",
-// 		yellow: "#E6E667CC",
-// 		white: "#D3D1EBD1",
-// 		red: "#E6457BCC",
-// 		brown: "#99611FCC",
-// 	},
-// }
+// Less Dimmed: {
+// 	pink: "#E68ECF",
+// 	magenta: "#C143AC",
+// 	purple: "#926EDB",
+// 	blue: "#429AD9",
+// 	cyan: "#42BDBD",
+// 	green: "#2AC264",
+// 	yellow: "#CFCF5D",
+// 	white: "#C0BED6",
+// 	red: "#D13E6C",
+// 	brown: "#85531C",
+// },
 
 const commentColors = {
-	"Dark Grey": "#737880",
-	"Light Grey": "#8A9099",
-	"Dark Magenta": "#9559B3",
-	"Light Magenta": "#AA66CC",
-	"Dark Forest": "#498C49",
-	"Light Forest": "#55A355",
-	"Dark Violet": "#6777CD",
-	"Light Violet": "#768AED",
-}
+	Normal: {
+		"Dark Grey": "#737880",
+		"Light Grey": "#8A9099",
+		"Dark Magenta": "#9559B3",
+		"Light Magenta": "#AA66CC",
+		"Dark Forest": "#498C49",
+		"Light Forest": "#55A355",
+		"Dark Violet": "#6777CD",
+		"Light Violet": "#768AED",
+	},
+	Dimmed: {
+		"Dark Grey": "#666B72",
+		"Light Grey": "#7B8088",
+		"Dark Magenta": "#854F9F",
+		"Light Magenta": "#975BB6",
+		"Dark Forest": "#417D41",
+		"Light Forest": "#4C914C",
+		"Dark Violet": "#5C6AB6",
+		"Light Violet": "#697AD3",
+	},
+};
 
 const schemes = (options, colors) => {
-	const schemes = {}
+	const schemes = {};
 	schemes[options.option_1] = {
 		string: colors.blue,
 		property: colors.magenta,
@@ -138,7 +76,7 @@ const schemes = (options, colors) => {
 		punctuation: colors.cyan,
 		invalid: colors.red,
 		deprecated: colors.brown,
-	}
+	};
 	schemes[options.option_2] = {
 		string: colors.green,
 		property: colors.purple,
@@ -150,7 +88,7 @@ const schemes = (options, colors) => {
 		punctuation: colors.cyan,
 		invalid: colors.red,
 		deprecated: colors.brown,
-	}
+	};
 	schemes[options.option_3] = {
 		string: colors.blue,
 		property: colors.magenta,
@@ -162,7 +100,7 @@ const schemes = (options, colors) => {
 		punctuation: colors.cyan,
 		invalid: colors.red,
 		deprecated: colors.brown,
-	}
+	};
 	schemes[options.option_4] = {
 		string: colors.green,
 		property: colors.purple,
@@ -174,14 +112,12 @@ const schemes = (options, colors) => {
 		punctuation: colors.cyan,
 		invalid: colors.red,
 		deprecated: colors.brown,
-	}
-	return schemes
-}
-
-const colorSchemes = schemes(nameOptions, syntaxColors) // dimmed
+	};
+	return schemes;
+};
 
 module.exports = {
 	syntaxColors,
-	colorSchemes,
 	commentColors,
-}
+	schemes,
+};