+++ src/vscode/user-preferences.js
@@ -1,24 +1,19 @@
-"use strict"
+"use strict";
 
-const {
-	themeTextName,
-	nameOptions
-} = require("./../../data/theme-names")
+const { themeTextName, nameOptions } = require("./../../data/theme-names");
 
-const {
-	themeMenu
-} = require("./menu-settings")
+const { themeMenu } = require("./menu-settings");
 
 // < gets saved configuration (settings.json)
-const settingsConfig = () => themeMenu().get("update.restoreConfiguration")
+const settingsConfig = () => themeMenu().get("update.restoreConfiguration");
 
-// ~ theme settings and favorite settings
-const themeConfig = () => settingsConfig().themeSettings
-const favoriteConfig = () => settingsConfig().favoriteSettings
+// > theme settings and favorite settings
+const themeConfig = () => settingsConfig().themeSettings;
+const favoriteConfig = () => settingsConfig().favoriteSettings;
 
 // < theme default configuration
 const themeDefault = (type) => {
-	let theme
+	let theme;
 	if (type === "theme") {
 		theme = {
 			background: "Grey",
@@ -26,9 +21,10 @@ const themeDefault = (type) => {
 			highlight: "Ultra Violet",
 			focusBorder: "ON",
 			fontStyle: "Italic",
+			syntaxBrightness: "Normal",
 			comments: "Dark Violet",
 			favorite: "OFF",
-		}
+		};
 	}
 	if (type === "favorite") {
 		theme = {
@@ -37,17 +33,18 @@ const themeDefault = (type) => {
 			highlight: "Ultra Violet",
 			focusBorder: "ON",
 			fontStyle: "Italic",
+			syntaxBrightness: "Dimmed",
 			comments: "Dark Magenta",
 			favorite: `${themeTextName} (${nameOptions.option_2})`,
-		}
+		};
 	}
-	return theme
-}
+	return theme;
+};
 
 // < checks for saved preferences:
-// ~ if "update.restoreConfiguration" is empty, it's a fresh install
+// > if "update.restoreConfiguration" is empty, it's a fresh install
 // ? if "update.restoreConfiguration" is has config, it's an update
-const freshInstall = Object.keys(settingsConfig()).length === 0
+const freshInstall = Object.keys(settingsConfig()).length === 0;
 
 // < creates the object that will be saved in the update.restoreConfig
 class Settings {
@@ -55,24 +52,24 @@ class Settings {
 		if (freshInstall) {
 			this.themeSettings = {
 				...themeDefault("theme"),
-			}
+			};
 			this.favoriteSettings = {
 				...themeDefault("favorite"),
-			}
+			};
 		} else if (config.favorite !== "OFF") {
 			this.themeSettings = {
 				...config,
-			}
+			};
 			this.favoriteSettings = {
 				...config,
-			}
+			};
 		} else {
 			this.themeSettings = {
 				...config,
-			}
+			};
 			this.favoriteSettings = {
 				...favoriteConfig(),
-			}
+			};
 		}
 	}
 }
@@ -83,14 +80,14 @@ const saveSettings = (config) => {
 		"update.restoreConfiguration",
 		new Settings(config),
 		true
-	)
-}
+	);
+};
 
 module.exports = {
 	favoriteConfig,
 	saveSettings,
 	themeConfig,
 	freshInstall,
-}
+};
 
 // syntaxBrightness: "Normal",